requirements.txt
absl-py==0.10.0
appdirs==1.4.4
astroid==2.4.2
astunparse==1.6.3
audioread==2.1.8
autopep8==1.5.4
azure-ai-textanalytics==5.1.0b1
azure-cognitiveservices-speech==1.13.0
azure-common==1.1.25
azure-core==1.8.2
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
decorator==4.4.2
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
isodate==0.6.0
isort==5.5.4
jedi==0.17.2
joblib==0.17.0
jupyter-client==6.1.7
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
librosa==0.8.0
llvmlite==0.34.0
Markdown==3.2.2
mccabe==0.6.1
msrest==0.6.19
numba==0.51.2
numexpr==2.7.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pooch==1.2.0
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
SoundFile==0.10.3.post1
tables==3.6.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.0
nltk>=3.3
sox>=1.3.3
